requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.32
pyodbc==5.1.0
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.2
email-validator==2.2.0
Jinja2==3.1.4